+++ src/Common/CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+Unless otherwise noted, all changes made by [@jeremyknight-me](https://github.com/jeremyknight-me).
+
 ## [Unreleased]
 
 ### Added
@@ -16,10 +18,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added AddWorkDays function to DateTimeHelper.
+- Added DoesImplement<T> to TypeExtensions.
 
 ### Changed
 - Updated email validation regex.
-- Upgraded .NET version from Core 2.1 to Core 3.0.
+- Upgraded .NET version to Core 3.0.
 
 [Unreleased]: https://github.com/jeremyknight-me/JK.Common/compare/1.0.6...HEAD
 [1.0.6]: https://github.com/jeremyknight-me/JK.Common/compare/1.0.5...1.0.6